pyproject.toml
[tool.poetry]
name = "todoapp"
version = "0.1.0"
description = "TODO with REST API"
authors = ["Claudio Souza"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.790"
pytest = "^6.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
